Remove backend-specific extension traits
These expose the implementations in the public API which prevents us
from upgrading versions or adjusting the specifics of our
implementation. It also kind of defeats the purpose of the crate itself.
They could come back in some form in the future, but we'd want to be
explicit that there are no stabilty guarantees.